refactor: some function to be independent in release commands
